splitting new_batches on several modules to reduce the module size
